refactor : sse 패키지 정리
